refs Ticket #2766
Increased maximum number of Joystick axes to 32.

NOTE: this will only result in more axes when the underlying hardware supports it; so no change in Windows; Linux probably will work (as it uses an ioctl to get the number from the driver), but this is untested.

Note sure what will happen on OSX.

ref: https://sourceforge.net/p/flightgear/codetickets/